Refactoring codebase for maintainability
