+++ src/main.rs
@@ -4,350 +4,17 @@ use anyhow::{Result, Context};
 use eframe::egui;
 use futures_util::StreamExt;
 use poll_promise::Promise;
-use serde::{Deserialize, Serialize};
 use reqwest::Client;
 use std::time::Duration;
 use std::sync::mpsc::{self, sync_channel, SyncSender};
 use tokio::runtime::Runtime;
+mod endpoint_type;
+mod llmclient;
+use endpoint_type::EndpointType;
+use llmclient::{LLMClient, ModelData, ModelsResponse, ChatMessage, ChatRequest, DeltaContent, Choice, ChatResponse, ModelDetails, OllamaModelsResponse};
 
 const DEFAULT_API_URL: &str = "http://localhost:1234/v1/chat/completions";
 const OLLAMA_API_URL: &str = "http://localhost:11434/v1/chat/completions";
-
-#[derive(Debug, Clone, PartialEq, Copy)]
-enum EndpointType {
-    LMStudio,
-    Ollama,
-}
-
-impl EndpointType {
-    fn default_url(&self) -> &'static str {
-        match self {
-            EndpointType::LMStudio => DEFAULT_API_URL,
-            EndpointType::Ollama => OLLAMA_API_URL,
-        }
-    }
-    
-    fn default_port(&self) -> &'static str {
-        match self {
-            EndpointType::LMStudio => "1234",
-            EndpointType::Ollama => "11434",
-        }
-    }
-
-    fn default_endpoint(&self) -> &'static str {
-        match self {
-            EndpointType::LMStudio => "v1/chat/completions",
-            EndpointType::Ollama => "v1/chat/completions",
-        }
-    }
-    
-    fn models_endpoint(&self, endpoint: &str) -> String {
-        match self {
-            EndpointType::LMStudio => {
-                // For LM Studio, always use /v1/models
-                "v1/models".to_string()
-            }
-            EndpointType::Ollama => {
-                // For Ollama, use /api/tags but respect any custom base path
-                if endpoint.is_empty() {
-                    "api/tags".to_string()
-                } else {
-                    // Strip the chat completions part if present and add api/tags
-                    let base = endpoint.trim_end_matches("v1/chat/completions");
-                    format!("{}api/tags", base.trim_end_matches('/'))
-                        .trim_start_matches('/')
-                        .to_string()
-                }
-            }
-        }
-    }
-
-    fn chat_endpoint(&self, endpoint: &str) -> String {
-        match self {
-            EndpointType::LMStudio => {
-                // For LM Studio, always use /v1/chat/completions
-                "v1/chat/completions".to_string()
-            }
-            EndpointType::Ollama => {
-                // For Ollama, use /api/chat but respect any custom base path
-                if endpoint.is_empty() {
-                    "api/chat".to_string()
-                } else {
-                    // Strip the chat completions part if present and add api/chat
-                    let base = endpoint.trim_end_matches("v1/chat/completions");
-                    format!("{}api/chat", base.trim_end_matches('/'))
-                        .trim_start_matches('/')
-                        .to_string()
-                }
-            }
-        }
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct ModelData {
-    id: String,
-    // Add other fields as needed
-}
-
-#[derive(Debug, Deserialize)]
-struct ModelsResponse {
-    data: Vec<ModelData>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct ChatMessage {
-    role: String,
-    content: String,
-}
-
-#[derive(Debug, Serialize)]
-struct ChatRequest {
-    model: String,
-    messages: Vec<ChatMessage>,
-    temperature: f32,
-    stream: bool,
-}
-
-#[derive(Debug, Deserialize)]
-struct DeltaContent {
-    content: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct Choice {
-    delta: DeltaContent,
-    finish_reason: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct ChatResponse {
-    choices: Vec<Choice>,
-}
-
-#[derive(Debug, Deserialize)]
-struct ModelDetails {
-    name: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct OllamaModelsResponse {
-    models: Vec<ModelDetails>,
-}
-
-#[derive(Clone)]
-struct LLMClient {
-    client: Client,
-    protocol: String,
-    server: String,
-    port: String,
-    endpoint: String,
-    endpoint_type: EndpointType,
-}
-
-impl LLMClient {
-    fn new(protocol: String, server: String, port: String, endpoint: String, endpoint_type: EndpointType) -> Self {
-        let client = Client::builder()
-            .timeout(Duration::from_secs(5))  // 5 second timeout
-            .build()
-            .unwrap_or_else(|_| Client::new());
-            
-        Self {
-            client,
-            protocol,
-            server,
-            port,
-            endpoint,
-            endpoint_type,
-        }
-    }
-
-    async fn list_models(&self) -> Result<Vec<String>> {
-        let models_url = format!("{}://{}:{}/{}",
-            self.protocol,
-            self.server,
-            self.port,
-            self.endpoint_type.models_endpoint(&self.endpoint)
-        ).trim_end_matches('/').to_string();
-        
-        println!("Fetching models from: {}", models_url);
-        
-        let response = self.client
-            .get(&models_url)
-            .send()
-            .await
-            .context(format!("Failed to fetch models from {}. Please check if the server is running and accessible", &models_url))?;
-            
-        if !response.status().is_success() {
-            return Err(anyhow::anyhow!(
-                "Server returned error {}: {}",
-                response.status(),
-                response.text().await.unwrap_or_else(|_| "No error message".to_string())
-            ));
-        }
-            
-        match self.endpoint_type {
-            EndpointType::LMStudio => {
-                let models: ModelsResponse = response
-                    .json()
-                    .await
-                    .context("Failed to parse models response")?;
-                    
-                Ok(models.data.into_iter().map(|m| m.id).collect())
-            }
-            EndpointType::Ollama => {
-                // First print the raw response for debugging
-                let text = response.text().await?;
-                println!("Raw Ollama response: {}", text);
-                
-                // Parse the response from the text
-                let models: OllamaModelsResponse = serde_json::from_str(&text)
-                    .context("Failed to parse Ollama response")?;
-                    
-                Ok(models.models.into_iter().map(|m| m.name).collect())
-            }
-        }
-    }
-
-    async fn chat_stream(&self, chat_history: &[(String, String)], prompt: &str, model: &str, tx: SyncSender<String>) -> Result<()> {
-        let chat_url = format!("{}://{}:{}/{}",
-            self.protocol,
-            self.server,
-            self.port,
-            self.endpoint_type.chat_endpoint(&self.endpoint)
-        ).trim_end_matches('/').to_string();
-        
-        // Convert chat history to messages format
-        let mut messages = Vec::new();
-        // Add all messages except the last one (which is the current prompt)
-        for (role, content) in chat_history.iter().take(chat_history.len().saturating_sub(1)) {
-            messages.push(serde_json::json!({
-                "role": role,
-                "content": content
-            }));
-        }
-        // Add current prompt
-        messages.push(serde_json::json!({
-            "role": "user",
-            "content": prompt
-        }));
-
-        // Different request format for different endpoints
-        let request_body = match self.endpoint_type {
-            EndpointType::LMStudio => {
-                let request = ChatRequest {
-                    model: model.to_string(),
-                    messages: messages.iter().map(|m| ChatMessage {
-                        role: m["role"].as_str().unwrap().to_string(),
-                        content: m["content"].as_str().unwrap().to_string(),
-                    }).collect(),
-                    temperature: 0.7,
-                    stream: true,
-                };
-                serde_json::to_value(request).unwrap()
-            }
-            EndpointType::Ollama => {
-                serde_json::json!({
-                    "model": model,
-                    "messages": messages,
-                    "stream": true
-                })
-            }
-        };
-
-        let response = self.client
-            .post(&chat_url)
-            .json(&request_body)
-            .timeout(Duration::from_secs(300))  // 5 minute timeout for the entire stream
-            .send()
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to send request: {}", e))?;
-
-        let status = response.status();
-        if !status.is_success() {
-            let error_text = response.text().await.unwrap_or_default();
-            return Err(anyhow::anyhow!("Request failed with status {}: {}", status, error_text));
-        }
-
-        let mut stream = response.bytes_stream();
-        let mut buffer = String::new();
-        let mut done = false;
-
-        while let Some(chunk_result) = stream.next().await {
-            match chunk_result {
-                Ok(chunk) => {
-                    let chunk_str = String::from_utf8_lossy(&chunk);
-                    
-                    match self.endpoint_type {
-                        EndpointType::LMStudio => {
-                            // LMStudio uses SSE format with "data: " prefix
-                            for line in chunk_str.lines() {
-                                let line = line.trim();
-                                if !line.starts_with("data: ") {
-                                    continue;
-                                }
-                                let line = &line["data: ".len()..];
-                                if line == "[DONE]" {
-                                    done = true;
-                                    continue;
-                                }
-
-                                if let Ok(response) = serde_json::from_str::<ChatResponse>(line) {
-                                    if let Some(choice) = response.choices.first() {
-                                        if let Some(content) = &choice.delta.content {
-                                            buffer.push_str(content);
-                                            tx.send(buffer.clone()).ok();
-                                        }
-                                        
-                                        if choice.finish_reason.is_some() {
-                                            done = true;
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                        EndpointType::Ollama => {
-                            // Ollama returns each chunk as a complete JSON object
-                            if let Ok(response) = serde_json::from_str::<serde_json::Value>(&chunk_str) {
-                                // Get content from message.content
-                                if let Some(message) = response.get("message") {
-                                    if let Some(content) = message.get("content") {
-                                        if let Some(text) = content.as_str() {
-                                            // Skip the thinking tokens but preserve newlines
-                                            if text != "<think>" && text != "</think>" {
-                                                // If we get pure newlines, add just one
-                                                if text.trim().is_empty() && text.contains('\n') {
-                                                    buffer.push('\n');
-                                                } else {
-                                                    buffer.push_str(text);
-                                                }
-                                                tx.send(buffer.clone()).ok();
-                                            }
-                                        }
-                                    }
-                                }
-                                
-                                if response.get("done").and_then(|v| v.as_bool()).unwrap_or(false) {
-                                    done = true;
-                                }
-                            }
-                        }
-                    }
-                }
-                Err(e) => {
-                    println!("Error reading chunk: {:?}", e);
-                    return Err(e).context("Failed to read chunk");
-                }
-            }
-        }
-
-        if done {
-            Ok(())
-        } else {
-            Err(anyhow::anyhow!("Stream ended without completion"))
-        }
-    }
-}
 
 struct ChatApp {
     client: LLMClient,
@@ -637,7 +304,7 @@ impl eframe::App for ChatApp {
         // Check for new response chunks
         if let Some(rx) = &self.response_receiver {
             if let Ok(new_content) = rx.try_recv() {
-                self.current_response = new_content;
+                self.current_response.push_str(&new_content);
                 ctx.request_repaint();
             }
         }
